Now supports inline eval using an echoing server.
This is a minimal example (or at least as small as I could make it)
that shows how to start a server from within LightTable and
communicate with it to provide ``evaluationg'' of selected
code.

For now the TCP server just echo's whatever you send to it but
it's a good first step if you want to develop real in-line
evaluation in LightTable.

To develop this code I had a look at the Python and Ruby
plugins.